script/requirements.txt
astroid==2.6.5
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
confluent-kafka==1.6.1
fastapi==0.65.0
Flask==1.1.2
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.0.0
h11==0.12.0
hiredis==2.0.0
httplib2==0.19.1
idna==3.2
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
neovim==0.3.1
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.1
pylint==2.9.5
pynvim==0.4.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rope==0.19.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
uvicorn==0.13.4
Werkzeug==1.0.1
wrapt==1.12.1
